requirements.txt
Flask
tensorflow-cpu
numpy
Pillow